Fixed prisma, new kubernetes file

+++ docker/cleanup.sh
@@ -1,5 +1,7 @@
 #!/bin/bash
 docker compose -f docker-compose.yml stop
 docker rm -f $(docker ps -a -q) dummy
-docker volume rm docker_db-data
+if [ "$1" -eq 1 ]; then
+    docker volume rm docker_db-data
+fi
 docker network prune -f
